rename src/page => src/ui
