+++ src/core/input_device.c
@@ -20,13 +20,13 @@
 #include "osd.h"
 #include "imagesetting.h"
 #include "common.hh"
-#include "../page/page_scannow.h"
-#include "../page/page_common.h"
-#include "../page/page_fans.h"
-#include "../page/page_power.h"
-#include "../page/page_imagesettings.h"
-#include "../page/page_playback.h"
-#include "../page/page_source.h"
+#include "ui/page_scannow.h"
+#include "ui/page_common.h"
+#include "ui/page_fans.h"
+#include "ui/page_power.h"
+#include "ui/page_imagesettings.h"
+#include "ui/page_playback.h"
+#include "ui/page_source.h"
 
 #include "../driver/oled.h"
 #include "../driver/dm6302.h"
