+++ src/core/main.c
@@ -28,9 +28,9 @@
 #include "../driver/mcp3021.h"
 #include "../driver/i2c.h"
 #include "../driver/esp32.h"
-#include "../page/page_scannow.h"
-#include "../page/page_power.h"
-#include "../page/page_source.h"
+#include "ui/page_scannow.h"
+#include "ui/page_power.h"
+#include "ui/page_source.h"
 #include "minIni.h"
 #include "../bmi270/accel_gyro.h"
 #include "thread.h"
