+++ src/core/main_menu.c
@@ -8,18 +8,18 @@
 #include "style.h"
 #include "porting.h"
 #include "common.hh"
-#include "../page/page_scannow.h"
-#include "../page/page_source.h"
-#include "../page/page_imagesettings.h"
-#include "../page/page_power.h"
-#include "../page/page_fans.h"
-#include "../page/page_record.h"
-#include "../page/page_autoscan.h"
-#include "../page/page_connections.h"
-#include "../page/page_headtracker.h"
-#include "../page/page_playback.h"
-#include "../page/page_version.h"
-#include "../page/page_common.h"
+#include "ui/page_scannow.h"
+#include "ui/page_source.h"
+#include "ui/page_imagesettings.h"
+#include "ui/page_power.h"
+#include "ui/page_fans.h"
+#include "ui/page_record.h"
+#include "ui/page_autoscan.h"
+#include "ui/page_connections.h"
+#include "ui/page_headtracker.h"
+#include "ui/page_playback.h"
+#include "ui/page_version.h"
+#include "ui/page_common.h"
 #include "../driver/hardware.h"
 #include "../driver/mcp3021.h"
 #include "../driver/oled.h"
