+++ src/core/osd.c
@@ -19,8 +19,8 @@
 #include "msp_displayport.h"
 #include "../driver/mcp3021.h"
 #include "../driver/nct75.h"
-#include "../page/page_common.h"
-#include "../page/page_fans.h"
+#include "ui/page_common.h"
+#include "ui/page_fans.h"
 #include "../driver/dm5680.h"
 #include "../driver/hardware.h"
 #include "fbtools.h"
