+++ src/core/thread.c
@@ -25,8 +25,8 @@
 #include "../driver/nct75.h"
 #include "../driver/oled.h"
 #include "../driver/porting.h"
-#include "../page/page_fans.h"
-#include "../page/page_version.h"
+#include "ui/page_fans.h"
+#include "ui/page_version.h"
 
 
 ///////////////////////////////////////////////////////////////////////////////
