+++ src/ui/page_common.c
