+++ src/ui/page_imagesettings.c
