+++ src/ui/page_power.c
