+++ src/ui/page_power.h
