+++ src/ui/page_scannow.c
